requirements.txt
pyyaml
networkx
pygraphviz
polars
